After fixing the cause of failure (Exception in thread "main" java.lang.IllegalArgumentException: PWC6309: Illegal compilerSourceVM: 1.7) in auto-complete.
This cause wasn't actually from the autocomplete. It was because, though the highest supported Java EE version of my NetBeans 7.1.2 is Java EE 6, I still wanted to code using Java 7 syntax. So I right-clicked project > properties > source/binary format > JDK 7

The actual cause of the error was because of this new settings. It doesn't permit me to run a JSP file. To circumvent that, I have to deploy the application and enter to URL of the JSP page.

Next step now is to use the auto-complete feature to this project's advantage. i.e. fetch records from db and display in auto-complete.